update install script

+++ put_dotfiles.sh
@@ -26,14 +26,15 @@ function ln_loop() {
     local org_dir=$1
     local org_files=`ls -1 $org_dir`
     for org_file in $org_files; do
-        ln -sfv $org_dir/$org_file ${HOME}/$file/$org_file
+        if [ ! -L ${HOME}/$file/$org_file ]
+            ln -sfv $org_dir/$org_file ${HOME}/$file/$org_file
+        fi
     done
 }
 
 for file in ${array[@]}; do
     echo $file
-    if [ -e ${HOME}/$file ]; then
-        echo "${HOME}/$file exists"
+    if [ -e ${HOME}/$file -a ! -L ${HOME}/$file ]; then
         mkdir -p ${HOME}/backup
         cp -pr ${HOME}/$file ${HOME}/backup/$file.$date
         # rm -rf ${HOME}/$file
